requirements.txt
faiss-cpu
sentence-transformers
langchain
streamlit
fastapi
pdfplumber
pymupdf
unstructured
python-dotenv
qdrant-client
chromadb
